requirements.txt
attrs==21.2.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2.13
django-import-export==2.5.0
django-rest-passwordreset==1.1.0
djangorestframework==3.12.4
drf-spectacular==0.17.2
et-xmlfile==1.1.0
inflection==0.5.1
jsonschema==3.2.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
pyrsistent==0.18.0
pytz==2018.5
PyYAML==5.4.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
ujson==4.0.2
uritemplate==3.0.1
xlrd==2.0.1
xlwt==1.3.0
